Move gpio common code to Button.c
