+++ STM32/Firmware/RelayBypass_STM32F101C4/Core/Src/gpio.c
@@ -21,10 +21,10 @@
 #include "gpio.h"
 
 /* USER CODE BEGIN 0 */
-bool gBtnStateA = false;
-bool gBtnStateB = false;
-
-static Node *pGpioList = NULL;
+//bool gBtnStateA = false;
+//bool gBtnStateB = false;
+//
+//static Node *pGpioList = NULL;
 /* USER CODE END 0 */
 
 /*----------------------------------------------------------------------------*/
@@ -121,52 +121,56 @@ void MX_GPIO_Init(void)
 /* USER CODE BEGIN 2 */
 void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin)
 {
-    HAL_TIM_Base_Start_IT(&htim2);
-    if ((GPIO_Pin & (A_BTN_Pin | B_BTN_Pin)) && ((gBtnStateA == false) || (gBtnStateB == false)))
-    {
-
-        if (GPIO_Pin & A_BTN_Pin)
-        {
-            gBtnStateA = true;
-        }
-
-        if (GPIO_Pin & B_BTN_Pin)
-        {
-            gBtnStateB = true;
-        }
-    }
+    Button_Callback(GPIO_Pin);
 }
-
-
-Status USER_GPIO_PushCommand(StateStruct *pCmd)
-{
-	if (pCmd == NULL)
-	{
-		return INVALID_PARAMETERS;
-	}
-
-	List_PushBack(&pGpioList, *pCmd);
-
-	return OK;
-}
-
-Status USER_GPIO_HandOverLocalList(Node **pMasterList)
-{
-    if (pMasterList == NULL)
-    {
-        return INVALID_PARAMETERS;
-    }
-
-	StateStruct temp;
-
-	while (pGpioList != NULL)
-	{
-	temp = List_Pop(&pGpioList);
-	List_PushBack(pMasterList, temp);
-	}
-
-	return OK;
-}
+//void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin)
+//{
+//    HAL_TIM_Base_Start_IT(&htim2);
+//    if ((GPIO_Pin & (A_BTN_Pin | B_BTN_Pin)) && ((gBtnStateA == false) || (gBtnStateB == false)))
+//    {
+//
+//        if (GPIO_Pin & A_BTN_Pin)
+//        {
+//            gBtnStateA = true;
+//        }
+//
+//        if (GPIO_Pin & B_BTN_Pin)
+//        {
+//            gBtnStateB = true;
+//        }
+//    }
+//}
+//
+//
+//Status USER_GPIO_PushCommand(StateStruct *pCmd)
+//{
+//	if (pCmd == NULL)
+//	{
+//		return INVALID_PARAMETERS;
+//	}
+//
+//	List_PushBack(&pGpioList, *pCmd);
+//
+//	return OK;
+//}
+//
+//Status USER_GPIO_HandOverLocalList(Node **pMasterList)
+//{
+//    if (pMasterList == NULL)
+//    {
+//        return INVALID_PARAMETERS;
+//    }
+//
+//	StateStruct temp;
+//
+//	while (pGpioList != NULL)
+//	{
+//	temp = List_Pop(&pGpioList);
+//	List_PushBack(pMasterList, temp);
+//	}
+//
+//	return OK;
+//}
 
 /* USER CODE END 2 */
