+++ STM32/Firmware/RelayBypass_STM32F101C4/STM32F101C4TX_FLASH.ld
@@ -37,7 +37,7 @@ _Min_Stack_Size = 0x400 ;	/* required amount of stack */
 MEMORY
 {
   RAM    (xrw)    : ORIGIN = 0x20000000,   LENGTH = 4K
-  FLASH    (rx)    : ORIGIN = 0x8001800,   LENGTH = 10K
+  FLASH    (rx)    : ORIGIN = 0x8000000,   LENGTH = 16K
 }
 
 /* Sections */
